FastAPI/requirements.txt
aiofiles==24.1.0
anyio==4.8.0
async-property==0.2.2
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.95.0
fastapi-keycloak-middleware==1.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jwcrypto==1.5.6
packaging==24.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==5.3.1
python-multipart==0.0.20
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.22.0
